feat: create shemas tool and system
