+++ services/agent_history/app/api/schemas/__init__.py
@@ -1,3 +1,5 @@
 from .agent import *
 from .discussion import *
 from .info import *
+from .system_message import *
+from .tool import *
